2024-d9p2: Massive misunderstanding. Just keeping intermediate solution
But actually this is all wrong. I was trying in the wrong direction. We
should try to move files once to the leftmost space that fits them, not
fill in the leftmost space with the files that fit. Even though I think
that this would be more compressed and efficient, it's not what the
problem statement wants lol.

The proof is in seeing that in the example, 7 is moved before 2.