2024d11-p2: Solution that can actually run in realistic time
As predicted, hashmap was the way.

Funnily enough, I managed to start the previous solution, in release
mode, develop this one, fix a small bug with the stone splitting, and
get the right output (in debug) before that one was even close to
completing (I assume - as I removed logging to make it faster).

Goes to show the power of optimization.