Simplify macro slightly
Since we now have the separate branch, we know we only enter this one if
we have at least part1. So we don't need to make that part optional and
can assume part1 is always given.

This also clears up weird cases where potentially we could provide part2
but not part1 values, which is not even really possible with the way AoC
works...